GameEngine - Change current approach of the engine to support clients with events notifications

+++ GameEngine/src/main/java/ies/portadaalta/gameengine/model/stats/CategoryStats.java
@@ -31,7 +31,7 @@ public class CategoryStats {
     public double getSuccessAnswerRate() {
         double rate = 0;
         if (numberOfQuestions != 0) {
-            rate = (rightAnswered / numberOfQuestions);
+            rate = (double)rightAnswered / numberOfQuestions;
         }
         return rate;
     }
